iOS/ARM64 port of FEXCore for jailed iOS devices
Port FEXCore to compile and run as a static library on iOS (arm64).
Key changes:

Build system:
- Skip Source/, Data/binfmts/, Data/AppConfig/ on Apple
- Disable jemalloc/rpmalloc, use system malloc
- Skip 64BitAllocator on Apple

Linux→Apple compatibility:
- prctl stubs, sigprocmask instead of SYS_rt_sigprocmask
- Spin-wait+usleep instead of futex (WritePriorityMutex, InterruptableConditionVariable)
- malloc_size instead of malloc_usable_size, posix_memalign instead of memalign
- sincos → sin+cos, fcopyfile instead of sendfile
- Relaxed InternalThreadState size assertions (Apple shared_mutex is larger)
- x18 excluded from register allocation (Apple platform TLS)

Dual-mapped JIT memory (W^X compliance):
- Buffer.h: WriteOffset/WritePtr for RX→RW translation
- JIT.cpp: SetWriteOffset(0) for TempCodeBuffer, restore for CodeBuffer
- Emitter.h: Forward label Bind() uses WritePtr for memory access
- Dispatcher.cpp: Split i-cache invalidation (data cache on RW, icache on RX)

Tested on iPhone 13 Pro (A15), iOS 26.3 — x86-64 exit(42) passes.

+++ FEXCore/Source/Interface/Core/CodeCache.cpp
@@ -303,7 +303,7 @@ bool CodeCache::SaveData(Core::InternalThreadState& Thread, int fd, const Execut
   char Zero[64] {};
   auto Off = lseek(fd, 0, SEEK_CUR);
   while (Off != AlignUp(Off, Utils::FEX_PAGE_SIZE)) {
-    auto BytesToWrite = std::min(AlignUp(Off, Utils::FEX_PAGE_SIZE) - Off, sizeof(Zero));
+    auto BytesToWrite = std::min(static_cast<size_t>(AlignUp(Off, Utils::FEX_PAGE_SIZE) - Off), sizeof(Zero));
     ::write(fd, Zero, BytesToWrite);
     Off += BytesToWrite;
   }
